refactor(store): extract custom SQL normalization
